Move api/global package to top level
