+++ exporters/trace/zipkin/zipkin.go
@@ -26,7 +26,7 @@ import (
 	"net/url"
 	"sync"
 
-	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/global"
 	export "go.opentelemetry.io/otel/sdk/export/trace"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 )
