+++ global/internal/meter_test.go
@@ -22,8 +22,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/api/global"
-	"go.opentelemetry.io/otel/api/global/internal"
+	"go.opentelemetry.io/otel/global"
+	"go.opentelemetry.io/otel/global/internal"
 	"go.opentelemetry.io/otel/label"
 	"go.opentelemetry.io/otel/metrictest"
 )
