+++ global/internal/registry_test.go
