+++ global/trace.go
@@ -15,8 +15,8 @@
 package global
 
 import (
-	"go.opentelemetry.io/otel/api/global/internal"
 	"go.opentelemetry.io/otel/api/trace"
+	"go.opentelemetry.io/otel/global/internal"
 )
 
 // Tracer creates a named tracer that implements Tracer interface.
