+++ global/trace_test.go
@@ -17,8 +17,8 @@ package global_test
 import (
 	"testing"
 
-	"go.opentelemetry.io/otel/api/global"
 	"go.opentelemetry.io/otel/api/trace"
+	"go.opentelemetry.io/otel/global"
 	"go.opentelemetry.io/otel/internal/trace/noop"
 )
