+++ propagators/trace_context_example_test.go
@@ -16,7 +16,7 @@ package propagators_test
 
 import (
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/global"
 	"go.opentelemetry.io/otel/propagators"
 )
