+++ sdk/trace/trace_test.go
@@ -25,7 +25,7 @@ import (
 	"testing"
 	"time"
 
-	"go.opentelemetry.io/otel/api/global"
+	"go.opentelemetry.io/otel/global"
 	"go.opentelemetry.io/otel/label"
 	"go.opentelemetry.io/otel/oteltest"
